+++ package.json
@@ -21,7 +21,7 @@
     "@types/react-dom": "18.0.9",
     "eslint": "8.30.0",
     "eslint-config-next": "13.0.7",
-    "firebase": "^9.15.0",
+    "firebase": "^9.14.0",
     "konva": "^8.3.14",
     "next": "13.0.7",
     "react": "18.2.0",
